linux color setter
