+++ src/main/kotlin/com/github/nowtilous/projectcolor/color_setters/ColorSetter.kt
@@ -2,12 +2,16 @@ package com.github.nowtilous.projectcolor.color_setters
 
 import com.github.nowtilous.projectcolor.gColorLockedComponentMap
 import com.github.nowtilous.projectcolor.gProjectColorMap
+import com.github.nowtilous.projectcolor.gProjectColorLockedMap
+import com.github.nowtilous.projectcolor.setTitleBarColor
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.WindowManager
 import com.intellij.ui.ColorUtil
 import java.awt.Color
 import java.awt.Component
 import java.awt.Container
+import java.awt.event.ContainerEvent
+import java.awt.event.ContainerListener
 import java.security.InvalidParameterException
 import javax.swing.JFrame
 
@@ -135,6 +139,34 @@ abstract class ColorSetter {
                     "foreground" -> component.foreground = if (ColorUtil.isDark(color)) Color.white else Color.black
                 }
             }
+        }
+    }
+
+    /**
+     * Lock color properties for given container of components.
+     *
+     * @param container the container to lock changes to.
+     * @param project the project to which the container belongs to.
+     *
+     * @note For some reason, Swing (or Intellij) completely swap out some container's components
+     *       when rebuilding the UI, so the color locking trick I did earlier is not enough,
+     *       since it will only lock components which might get disposed of some day...
+     *       Hence, this function is introduced, which adds a listener to container change events
+     *       such as component added, component removed, etc.
+     *       Once such event occurs, the title bar color will be reset.
+     */
+    fun lockContainerColor(container: Container, project: Project) {
+        if (!gProjectColorLockedMap.containsKey(project) || gProjectColorLockedMap[project] == false) {
+            container.addContainerListener(object : ContainerListener {
+                override fun componentAdded(e: ContainerEvent?) {
+                    gProjectColorMap[project]?.let { setTitleBarColor(it, project) }
+                }
+
+                override fun componentRemoved(e: ContainerEvent?) {
+                    gProjectColorMap[project]?.let { setTitleBarColor(it, project) }
+                }
+            })
+            gProjectColorLockedMap[project] = true
         }
     }
 }
